+++ ampligraph/latent_features/models/ComplEx.py
@@ -187,14 +187,22 @@ class ComplEx(EmbeddingModel):
         """
         if not self.dealing_with_large_graphs:
             self.ent_emb = tf.get_variable('ent_emb', shape=[len(self.ent_to_idx), self.internal_k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=self.initializer.get_entity_initializer(
+                                           len(self.ent_to_idx), self.internal_k),
+                                           dtype=tf.float32)
             self.rel_emb = tf.get_variable('rel_emb', shape=[len(self.rel_to_idx), self.internal_k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=self.initializer.get_relation_initializer(
+                                           len(self.rel_to_idx), self.internal_k),
+                                           dtype=tf.float32)
         else:
+            # initialize entity embeddings to zero (these are reinitialized every batch by batch embeddings)
             self.ent_emb = tf.get_variable('ent_emb', shape=[self.batch_size * 2, self.internal_k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=tf.zeros_initializer(),
+                                           dtype=tf.float32)
             self.rel_emb = tf.get_variable('rel_emb', shape=[len(self.rel_to_idx), self.internal_k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=self.initializer.get_relation_initializer(
+                                           len(self.rel_to_idx), self.internal_k),
+                                           dtype=tf.float32)
 
     def _fn(self, e_s, e_p, e_o):
         r"""ComplEx scoring function.