+++ ampligraph/latent_features/models/ConvKB.py
@@ -1,4 +1,4 @@
-# Copyright 2019 The AmpliGraph Authors. All Rights Reserved.
+# Copyright 2019-2020 The AmpliGraph Authors. All Rights Reserved.
 #
 # This file is Licensed under the Apache License, Version 2.0.
 # A copy of the Licence is available in LICENCE, or at:
@@ -200,17 +200,20 @@ class ConvKB(EmbeddingModel):
         if not self.dealing_with_large_graphs:
 
             self.ent_emb = tf.get_variable('ent_emb', shape=[len(self.ent_to_idx), self.k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=self.initializer.get_entity_initializer(
+                                           len(self.ent_to_idx), self.k), dtype=tf.float32)
             self.rel_emb = tf.get_variable('rel_emb', shape=[len(self.rel_to_idx), self.k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=self.initializer.get_relation_initializer(
+                                           len(self.rel_to_idx), self.k), dtype=tf.float32)
 
         else:
 
             self.ent_emb = tf.get_variable('ent_emb', shape=[self.batch_size * 2, self.internal_k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=tf.zeros_initializer(), dtype=tf.float32)
 
             self.rel_emb = tf.get_variable('rel_emb', shape=[len(self.rel_to_idx), self.internal_k],
-                                           initializer=self.initializer.get_tf_initializer(), dtype=tf.float32)
+                                           initializer=self.initializer.get_relation_initializer(
+                                           len(self.rel_to_idx), self.internal_k), dtype=tf.float32)
 
         num_filters = self.embedding_model_params['num_filters']
         filter_sizes = self.embedding_model_params['filter_sizes']