+++ ampligraph/latent_features/models/TransE.py
@@ -3,6 +3,7 @@ from ampligraph.latent_features import constants as constants
 from ampligraph.latent_features.initializers import DEFAULT_XAVIER_IS_UNIFORM
 import tensorflow as tf
 
+
 @register_model("TransE",
                 ["norm", "normalize_ent_emb", "negative_corruption_entities"])
 class TransE(EmbeddingModel):